hooks: split playback controller
