+++ crates/hooks/src/lib.rs
@@ -4,6 +4,8 @@
 pub mod db_reactivity;
 pub mod debug_db;
 pub mod playback_ref;
+mod player_controller_queue;
+pub mod scrobble_scheduler;
 pub mod source_switch;
 pub mod use_db_queries;
 pub mod use_player_controller;
